+++ packages/arm-core/src/components/ControlButton.ts
@@ -1,5 +1,5 @@
+import { Button } from '@amsterdam/asc-ui'
 import styled from 'styled-components'
-import { Button } from '@datapunt/asc-ui'
 
 const ControlButton = styled(Button)<any>`
   box-shadow: 0 0 0 2px rgba(0, 0, 0, 0.1);