+++ packages/arm-core/src/components/MapPanel/MapPanelLegendButton.tsx
@@ -1,5 +1,5 @@
+import { MapLayers } from '@amsterdam/asc-assets'
 import React, { useContext } from 'react'
-import { MapLayers } from '@datapunt/asc-assets'
 import ControlButton from '../ControlButton'
 import { Overlay, SnapPoint } from './constants'
 import MapPanelContext from './MapPanelContext'