+++ packages/arm-core/src/components/Scale.tsx
@@ -1,4 +1,4 @@
-import { useMapInstance } from '@datapunt/react-maps'
+import { useMapInstance } from '@amsterdam/react-maps'
 import { Control, control } from 'leaflet'
 import React, { useEffect } from 'react'