+++ packages/arm-core/src/utils/usePanToLatLng.ts
@@ -1,4 +1,4 @@
-import { useMapInstance } from '@datapunt/react-maps'
+import { useMapInstance } from '@amsterdam/react-maps'
 import { LatLngLiteral } from 'leaflet'
 
 const usePanToLatLng = () => {